Move wiremock stubs out of test directory
We'll be using these stubs in the local and development environments
while we don't have an exisiting endpoint available in the API, so
it no longer makes sense for them to live inside a test-focused
directory
